feat(prometheus): allow to convert string values from sensors into values
Currently JTImon would omit string values, but for certain sensors exposed by JTI
it would be great to have access to these values.

In order to do so, we export them as a constat metric with value 1 and put
the string value into a label. In order to not put all strings into labels
and risk cardinality churn for prometheus, a new config has been added

Example:
{
   "prometheus": {
      "string-as-label": ["/network-instances/network-instance/protocols/protocol/bgp/neighbors/neighbor/state/session-state"]
   },
}